feat: improve keyframes scoring by using a weighted average of the full score with the maximun of the bodypart score

+++ apps/ergo4all/lib/common/rula_session.dart
@@ -57,8 +57,8 @@ class KeyFrame {
   /// KeyFrame init
   const KeyFrame(this.score, this.screenshot, this.timestamp);
 
-  /// KeyFrame full score
-  final int score;
+  /// KeyFrame score
+  final double score;
 
   /// keyFrame screenshot
   final Uint8List screenshot;
